# [Enhancement] Optimize the memory usage of rows between unbounded preceding and current row for non-pipeline engine